moved page view into respective pages folder
Relocated all pages into their own folder. All imports have been updated accordingly.
